+++ examples/development/variables.tf
@@ -1,6 +1,6 @@
 variable "subscription_id" { type = string }
 variable "tenant_id" { type = string }
-variable "cluster_name" { type = string }
+variable "naming" { type = any }
 variable "resource_group" { type = any }
 variable "node_subnet_id" { type = string }
 variable "kubernetes_version" {
@@ -19,6 +19,7 @@ variable "autoscaling" { type = any }
 variable "disruption_profile" { type = any }
 variable "private_cluster" { type = any }
 variable "network" { type = any }
+variable "kms_encryption" { type = any }
 variable "system_node_pool" { type = any }
 variable "user_node_pools" { type = any }
 variable "auto_scaler_profile" { type = any }
@@ -32,3 +33,19 @@ variable "backup_integration" { type = any }
 variable "pki_integration" { type = any }
 variable "diagnostic_log_categories" { type = set(string) }
 variable "tags" { type = map(string) }
+variable "manage_role_assignments" {
+  description = "Whether this module creates Azure RBAC role assignments required by AKS."
+  type        = bool
+}
+variable "managed_identities" {
+  description = "Existing or module-managed AKS identities."
+  type = object({
+    create                    = bool
+    control_plane_identity_id = optional(string)
+    kubelet_identity_id       = optional(string)
+  })
+}
+# variable "location" {
+#   description = "Azure region in which regional AKS resources are deployed."
+#   type        = string
+# }